requirements.txt
django==4.2.6
django-rest-framework==0.1.0
djangorestframework~=3.14.0
